+++ eGym.Core.Domain/AggregateModels/Todo/TodoSpecifications.cs
@@ -12,7 +12,7 @@ namespace eGym.Core.Domain
         public static ASpec<Todo_Master> ByPriorities(params TodoPriorityEnum[] priorities) => new Spec<Todo_Master>(t => (priorities == null || priorities.Count() <= 0) || priorities.Contains(t.TD_Priority));
 
         public static ASpec<Todo_Master> ByStatusID(int id) => new Spec<Todo_Master>(t => t.TD_StatusID == id);
-        public static ASpec<Todo_Master> ByStatusIDs(params int[] ids) => new Spec<Todo_Master>(t => (ids == null || ids.Count() <= 0) || ids.Contains(t.TD_StatusID));
+        public static ASpec<Todo_Master> ByStatusIDs(params short[] ids) => new Spec<Todo_Master>(t => (ids == null || ids.Count() <= 0) || ids.Contains(t.TD_StatusID));
 
         public static ASpec<Todo_Master> BySearchString(string search)
             => new Spec<Todo_Master>(t => string.IsNullOrWhiteSpace(search) || t.TD_Title.Contains(search.Trim()) /*|| t.TD_Content.Contains(search.Trim())*/);
