+++ eGym.MVC/Areas/Admin/Models/TodoFilter.cs
@@ -8,7 +8,7 @@ namespace eGym.MVC.Areas.Admin.Models
 {
     public class TodoFilter
     {
-        public int[] Statuses { get; set; } = EN_TodoStatus.GetAll().Select(t => (int)t.ID).ToArray();
+        public short[] Statuses { get; set; } = EN_TodoStatus.GetAll().Select(t => t.ID).ToArray();
         public string SearchString { get; set; } = null;
         public bool Important { get; set; } = false;
     }
